Address PR feedback comments
- Change internal delay types to uint16_t (proto remains uint32)
- Remove redundant deactivation work cancel in input handler
- Deactivate immediately without delay in keep_active callback
- Use keycode_state_changed event for timestamp tracking
- Use position_state_changed for deactivation with proper binding checks
- Check auto-mouse layer binding - don't deactivate if non-transparent
- Check resolved binding - don't deactivate for modifier keys
- All tests passing

+++ include/zmk/pointing/input_processor_runtime.h
@@ -19,8 +19,8 @@ struct zmk_input_processor_runtime_config {
     // Auto-mouse layer settings
     bool auto_mouse_enabled;
     uint8_t auto_mouse_layer;
-    uint32_t auto_mouse_activation_delay_ms;
-    uint32_t auto_mouse_deactivation_delay_ms;
+    uint16_t auto_mouse_activation_delay_ms;
+    uint16_t auto_mouse_deactivation_delay_ms;
 };
 
 /**
